twilio-cmdline/requirements.txt
twilio~=7.15.3
python-dotenv~=0.21.0